requirements-dev.txt
-r requirements.txt
flake8==3.5.0
mock==1.0.1
pytest==3.2.3
requests-mock==0.6.0
